Views for teacher (#21)
* solve merge conflicts manually

* create language page for adding teachers

* fix merge conflicts in views

* resolve rebase conflict in views

* begin to implement pages to add and administrate teachers

* finish add form for teachers, add multiple course select field to cached

* add basic db.model for teacher

* add edit_teacher view, finish add_teacher function, implement list to show all teachers of one language, finish db.Model Teacher

* implement logic to add/remove courses to teachers, delete Teacher attribute languages (implicit known through link between courses and languages -> be efficient), course is only addable to a teacher, if it is not assigned already

* put administration view functions in admin_views (own package), code view for teachers to manage their courses (def teacher in views)

* add attendance base view and view to put the grade

* add grade edit views and intern representation for the grade

* changes to insert grades (at the moment not fully working)

* export for participation certificates (#16)

* add check boxes to courses.html, implement participation certificate generator

* return applicant mail if checkbox in course.html is marked

* write logic for participation_cert pdf generation and implement zip-writer

* nicer checkbox design

* minor final changes

* add teacher administration view

* Merge new User database model (#18)

* update excel templates

* export for participation certificates (#16)

* add check boxes to courses.html, implement participation certificate generator

* return applicant mail if checkbox in course.html is marked

* write logic for participation_cert pdf generation and implement zip-writer

* nicer checkbox design

* minor final changes

* minor changes in spanish templates

* bugfix on spanisch_hueber

* give all users the possibility to generate participation certificates

* Fix non-discounted price being shown in outstanding payments

* Use roles to store user permissions

* Add helper to check if user has superuser role

* Rename LANGUAGE_ADMIN role to COURSE_ADMIN

* Update can_edit_course check with new role system

* Create log entries for courses, not for languages

* Replace usages of user.languages

* Add additional teacher attributes to user model

---------


* implementation of new database model for teacher administration views

* finish integration of database model, adapt edit_teacher with add/remove course

* prefill edit grade field and show corresponding field in template

* show statistics for teacher administration, code attendance table

* adjust design and spacing of attendance table

* attempt using Flask WTForms for the input handling of the grade (edit_grade.html). Unfortunately the FieldList with FormField causes problems e g Field Validation seems not to work, flashed messages not displayed,...

* implement edit_grade form to adjust ects points for each student and set the grade in percentages, later it is shown as grade with comma

* add hide_grade template

* finish implementation of the templates for setting the grade, ects and if the grade should be hidden

* fix logic error in edit_grade

* adapt model to new database role

* send pw reset mails to teachers, display teachers in course template, begin to limit views for teachers

* show courses that are not assigned to a teacher in teacher overview

* fix course assigning to teachers, make error messages work as intended

* show only avalable courses in add_teacher form

* restrict access for teachers to only necessary pages

* integrate edit grade flow into language table of courses as own column

* fix grade routing for admins, add page for teachers without any course to make new course assignments possible

* fix teacher routing, remove complete row in database when removing course role, fix check availability function to assign new courses to a teacher

* minor view enhancements

---------

+++ src/spz/views.py
@@ -3,6 +3,8 @@
 """The application's views.
 
    Manages the mapping between routes and their activities.
+
+   Notice: administration_views contains the views for teacher administration
 """
 
 import socket
@@ -25,6 +27,7 @@ import spz.forms as forms
 from spz.util.Filetype import mime_from_filepointer
 from spz.mail import generate_status_mail
 from spz.export import export_course_list
+from spz.administration import TeacherManagement
 
 from flask_babel import gettext as _
 
@@ -33,6 +36,8 @@ from spz.oidc import oidc_callback, oidc_url, oidc_get_resources
 from spz.pdf_zip import PdfZipWriter, html_response
 from spz.pdf import generate_participation_cert
 from spz.auth.password_reset import validate_reset_token_and_get_user_id
+
+from spz.administration import TeacherManagement
 
 
 def check_precondition_with_auth(cond, msg, auth=False):
@@ -312,6 +317,8 @@ def signupinternal(course_id):
                 discount=applicant.current_discount(),
                 informed_about_rejection=informed_about_rejection
             )
+            # set course ects to applicant
+            applicant.ects_points = course.ects_points
             db.session.add(applicant)
             db.session.delete(o_auth_token)
             db.session.commit()
@@ -426,6 +433,8 @@ def signupexternal(course_id):
                 discount=applicant.current_discount(),
                 informed_about_rejection=informed_about_rejection
             )
+            # add course ects to applicant
+            applicant.ects_points = course.ects_points
             db.session.add(applicant)
             db.session.commit()
         except Exception as e:
@@ -501,6 +510,8 @@ def signoff():
 @login_required
 @templated('internal/overview.html')
 def internal():
+    if current_user.is_teacher:
+        return redirect(url_for('teacher'))
     logs = models.LogEntry.get_visible_log(current_user, 200)
     return dict(logs=logs)
 
@@ -508,6 +519,8 @@ def internal():
 @login_required
 @templated('internal/registrations.html')
 def registrations():
+    if current_user.is_teacher:
+        return redirect(url_for('teacher'))
     form = forms.TagForm()
     return dict(form=form)
 
@@ -576,6 +589,8 @@ def registrations_verify():
 @login_required
 @templated('internal/approvals.html')
 def approvals():
+    if current_user.is_teacher:
+        return redirect(url_for('teacher'))
     form = forms.TagForm()
     return dict(form=form)
 
@@ -690,6 +705,8 @@ def approvals_check():
 @login_required
 @templated('internal/notifications.html')
 def notifications():
+    if current_user.is_teacher:
+        return redirect(url_for('teacher'))
     form = forms.NotificationForm()
 
     if form.validate_on_submit():
@@ -760,12 +777,18 @@ def export(type, id):
             language = models.Language.query.get(id)
             form.courses.data = [course.id for course in language.courses]
 
-    return dict(form=form)
+    # update course multi-select based on assigned courses to user
+    # if teacher, then only own courses in multi-select selectable
+    # if superuser or course admin, then all courses can be downloaded
+    form.update_course_list(current_user)
+    return dict(form=form, user=current_user)
 
 
 @login_required
 @templated('internal/lists.html')
 def lists():
+    if current_user.is_teacher:
+        return redirect(url_for('teacher'))
     # list of tuple (lang, aggregated number of courses, aggregated number of seats)
     lang_misc = db.session.query(models.Language, func.count(models.Language.courses), func.sum(models.Course.limit)) \
         .join(models.Course, models.Language.courses) \
@@ -779,15 +802,33 @@ def lists():
 @login_required
 @templated('internal/language.html')
 def language(id):
+    if current_user.is_teacher:
+        return redirect(url_for('teacher'))
     return dict(language=models.Language.query.get_or_404(id))
 
 
 @login_required
 @templated('internal/course.html')
 def course(id):
+    if current_user.is_teacher:
+        return redirect(url_for('teacher'))
     course = models.Course.query.get_or_404(id)
     form = forms.CourseForm()
     form_delete = forms.DeleteCourseForm()
+
+    # check if a teacher has been assigned to the course
+    teacher = (
+        models.User.query
+        .join(models.Role, models.User.roles)
+        .filter(models.Role.course_id == course.id)
+        .filter(models.Role.role == 'COURSE_TEACHER')
+        .distinct()
+        .all()
+    )
+    if len(teacher) > 1:
+        flash(
+            _('Achtung: Der Kurs hat mehr als nur einen Dozenten zugewiesen. Das ist ungültig',),
+            'error')
 
     # we have two forms on this page, to differ between them a hidden identifier tag is used
 
@@ -854,7 +895,7 @@ def course(id):
                 _('Der Kurs konnte nicht gelöscht werden: %(error)s', error=e),
                 'error'
             )
-    return dict(course=course, form=form, form_delete=form_delete)
+    return dict(course=course, form=form, form_delete=form_delete, teacher=teacher)
 
 
 @login_required
@@ -924,6 +965,8 @@ def applicant(id):
 @login_required
 @templated('internal/applicants/search_applicant.html')
 def search_applicant():
+    if current_user.is_teacher:
+        return redirect(url_for('teacher'))
     form = forms.SearchForm()
 
     applicants = []
@@ -1013,6 +1056,8 @@ def applicant_attendances(id):
 @login_required
 @templated('internal/payments.html')
 def payments():
+    if current_user.is_teacher:
+        return redirect(url_for('teacher'))
     form = forms.PaymentForm()
 
     if form.validate_on_submit():
@@ -1043,6 +1088,8 @@ def payments():
 @login_required
 @templated('internal/outstanding.html')
 def outstanding():
+    if current_user.is_teacher:
+        return redirect(url_for('teacher'))
     outstanding = db.session.query(models.Attendance) \
         .join(models.Course, models.Applicant) \
         .filter(not_(models.Attendance.waiting),
@@ -1089,6 +1136,8 @@ def status(applicant_id, course_id):
 @login_required
 @templated('internal/statistics.html')
 def statistics():
+    if current_user.is_teacher:
+        return redirect(url_for('teacher'))
     return None
 
 
@@ -1138,6 +1187,8 @@ def task_queue():
 @login_required
 @templated('internal/preterm.html')
 def preterm():
+    if current_user.is_teacher:
+        return redirect(url_for('teacher'))
     form = forms.PretermForm()
 
     token = None
@@ -1173,6 +1224,8 @@ def preterm():
 @login_required
 @templated('internal/duplicates.html')
 def duplicates():
+    if current_user.is_teacher:
+        return redirect(url_for('teacher'))
     taglist = db.session.query(models.Applicant.tag) \
         .filter(models.Applicant.tag is not None, models.Applicant.tag != '') \
         .group_by(models.Applicant.tag) \
@@ -1230,6 +1283,8 @@ def login():
         user = models.User.get_by_login(form.user.data, form.password.data)
         if user:
             login_user(user, remember=True)
+            if current_user.is_teacher:
+                return redirect(url_for('teacher'))
             return redirect(url_for('internal'))
         flash(_('Du kommst hier net rein!'), 'negative')
 
@@ -1250,14 +1305,16 @@ def reset_password(reset_token):
         userId = validate_reset_token_and_get_user_id(reset_token)
 
         if userId is False:
-            flash(_('Das Passwort konnte nicht festgelegt werden. Bitte konraktieren Sie das Sprachenzentrum.'), 'negative')
+            flash(_('Das Passwort konnte nicht festgelegt werden. Bitte konraktieren Sie das Sprachenzentrum.'),
+                  'negative')
 
             return dict(form=form)
 
         user = models.User.query.filter(models.User.id == userId).first()
 
         if not user:
-            flash(_('Das Passwort konnte nicht festgelegt werden. Bitte konraktieren Sie das Sprachenzentrum.'), 'negative')
+            flash(_('Das Passwort konnte nicht festgelegt werden. Bitte konraktieren Sie das Sprachenzentrum.'),
+                  'negative')
 
             return dict(form=form)
